app/requirements.txt
django==4.0.4
uwsgi==2.0.20
flake8==4.0
python-dotenv==0.6.4
psycopg2==2.9.3
django-debug-toolbar==3.4.0
django-split-settings==1.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.8.0
pytz==2022.1